# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Espaço Sinapse — API administrativa",
    "description": "Sessão HttpOnly. Obtenha GET /api/v1/auth/csrf e envie seu token no headerName retornado em operações de escrita. Após login, obtenha novo CSRF. ADMIN administra cadastros; RECEPCAO gerencia pacientes e agenda. Datas usam offset ISO-8601. Atualizações exigem version. Erros usam ProblemDetail; conflitos retornam 409, incluindo conflictingAppointmentIds quando aplicável. API pública expõe somente campos aprovados.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/password": {
      "put": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "password",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "description": "Entrada pública por e-mail e senha; não exige sessão previamente autenticada. Antes de enviar, consulte GET /api/v1/auth/csrf, conserve o cookie de sessão anônima e envie X-CSRF-TOKEN. Após sucesso, a sessão fica autenticada e o CSRF anterior é invalidado; consulte /auth/csrf novamente. CSRF ausente ou inválido retorna 403; credenciais inválidas, 401; limite de tentativas excedido, 429.",
        "operationId": "login",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthDto"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/appointments": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "list_4",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "professionalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "patientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AGENDADO",
                "CONFIRMADO",
                "CONCLUIDO",
                "CANCELADO",
                "NAO_COMPARECEU"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoAppointmentDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "create_4",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/appointments/{id}/status": {
      "patch": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{id}/reschedule": {
      "patch": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "reschedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/appointments/{id}": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "get_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{id}/history": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasswordInput": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 72,
            "minLength": 0
          },
          "newPassword": {
            "type": "string",
            "maxLength": 72,
            "minLength": 12
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 0
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "RECEPCAO"
            ]
          }
        }
      },
      "AppointmentInput": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "professionalId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "patientId",
          "professionalId",
          "serviceId",
          "start"
        ]
      },
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "patientName": {
            "type": "string"
          },
          "professionalId": {
            "type": "string",
            "format": "uuid"
          },
          "professionalName": {
            "type": "string"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceName": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "AGENDADO",
              "CONFIRMADO",
              "CONCLUIDO",
              "CANCELADO",
              "NAO_COMPARECEU"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StatusInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AGENDADO",
              "CONFIRMADO",
              "CONCLUIDO",
              "CANCELADO",
              "NAO_COMPARECEU"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "status",
          "version"
        ]
      },
      "RescheduleInput": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "start",
          "version"
        ]
      },
...
      "PageDtoAppointmentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentDto"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "HistoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "previousStart": {
            "type": "string",
            "format": "date-time"
          },
          "newStart": {
            "type": "string",
            "format": "date-time"
          },
          "previousStatus": {
            "type": "string"
          },
          "newStatus": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}